backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
langchain-perplexity
langchain-experimental
pypdf
python-docx
chromadb
sentence-transformers
doc2txt